#214 fix: 컨플릭트 수정
