+++ src/main/java/com/floney/floney/book/repository/BookUserCustomRepository.java
@@ -35,9 +35,7 @@ public interface BookUserCustomRepository {
 
     boolean existsByUserEmailAndBookKey(String email, String bookKey);
 
-    Optional<User> findBookUserWhoSubscribeExclusively(Book targetBook);
-
-    Optional<User> findRandomBookUserWhoSubscribe(Book targetBook);
+    Optional<User> findRandomBookUserWhoSubscribeExclusively(Book targetBook);
 
     List<BookUser> findAllByUserId(Long userId);
