+++ src/main/java/com/floney/floney/book/service/BookLineService.java
@@ -22,5 +22,5 @@ public interface BookLineService {
 
     BookLineResponse changeLine(ChangeBookLineRequest request);
 
-    void deleteLine(Long bookLineKey);
+    void deleteLine(Long bookLineId);
 }
